Add basic support for NPAPI downloading of files.  Implements NPN_GetURL and includes a basic test showing the cross-origin checking implemented.
Review URL: http://codereview.chromium.org/570010

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@1468 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2

Former-commit-id: a69a1c250385c6d62b0856d87bd46807ef9bcba8

+++ src/shared/npruntime/device2d.cc
@@ -110,14 +110,16 @@ cleanup:
 static NPError GetStateContext(NPP instance,
                                NPDeviceContext* context,
                                int32 state,
-                               int32 *value) {
+                               intptr_t *value) {
+  // TODO(sehr): Add an RPC for this when 2D trusted API implements it.
   return NPERR_GENERIC_ERROR;
 }
 
 static NPError SetStateContext(NPP instance,
                                NPDeviceContext* context,
                                int32 state,
-                               int32 value) {
+                               intptr_t value) {
+  // TODO(sehr): Add an RPC for this when 2D trusted API implements it.
   return NPERR_GENERIC_ERROR;
 }
